+++ src/main/java/com/osoondosson/controller/ImageUploadController.java
@@ -2,16 +2,15 @@ package com.osoondosson.controller;
 
 import java.security.Principal;
 
+import com.osoondosson.service.S3FileService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.osoondosson.service.ImageUploadService;
 import com.osoondosson.service.UserService;
 import com.osoondosson.vo.UploadResponse;
 import com.osoondosson.vo.UserVO;
@@ -22,19 +21,19 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class ImageUploadController {
 
-	private final ImageUploadService imageUploadService;
+	private final S3FileService s3FileService;
 	private final UserService userService;
 
     @Autowired
-    public ImageUploadController(ImageUploadService imageUploadService, UserService userService) {
-        this.imageUploadService = imageUploadService;
+    public ImageUploadController(S3FileService s3FileService, UserService userService) {
+        this.s3FileService = s3FileService;
         this.userService = userService;
     }
 
     @PostMapping("/student/fileUpload" )
     public ResponseEntity<UploadResponse> uploadImage(@RequestParam("fileData") MultipartFile file) {
         try {
-            String result = imageUploadService.uploadFile(file);
+            String result = s3FileService.uploadFile(file);
             return ResponseEntity.ok(new UploadResponse("success",result));
         } catch (Exception e) {
             log.error("Error occurred while uploading file", e);
@@ -47,7 +46,7 @@ public class ImageUploadController {
     public ResponseEntity<UploadResponse> uploadImage2(@RequestParam("fileData") MultipartFile file, Principal principal) {
         try {
         	String userId = principal.getName();
-            String result = imageUploadService.uploadFile2(file, userId);
+            String result = s3FileService.uploadFile2(file, userId);
             return ResponseEntity.ok(new UploadResponse("success",result));
         } catch (Exception e) {
             log.error("Error occurred while uploading file", e);
@@ -61,7 +60,7 @@ public class ImageUploadController {
         try {
 
         	UserVO vo = userService.findById(principal.getName());
-        	imageUploadService.deleteFile(vo);
+        	s3FileService.deleteFile(vo);
             return ResponseEntity.noContent().build();
         } catch (IllegalArgumentException e) {
             return ResponseEntity.badRequest().body(new UploadResponse("error","No file associated with this user"));
