+++ src/main/resources/mapper/FileMapper.xml
@@ -31,7 +31,7 @@
     <insert id="insertImage" parameterType="fileVO">
 		insert into FILETABLE VALUES(#{uuid}, #{originName}, #{uploadName}, #{uploadPath}, #{fileType}, CURRENT_TIMESTAMP, 1, GROUP_SEQ.nextval, #{fileSize})
 	</insert>
-	
+
 	<delete id="deleteImage" parameterType="fileVO">
 		delete from FILETABLE where uuid = #{uuid}
 	</delete>
